3.Recurrent Neural Networks/requirements.txt
numpy==1.22.1
wandb==0.12.10
tqdm
torch
torch_xla
tensorflow-gpu
tensorflow-addons
matplotlib
transformers 
datasets 
accelerate
